Start Window class, to represent a terminal window
Started out as copypasta from the current View code. Window is meant to be an
OOP wrapper around ncurses, as much as possible. Currently, including Window
polllutes the global namespace with ncurses things, but that will be fixed
eventually. Offers basic functionality like setting up curses mode and get the
next key pressed.

The .hpp contains a Key namespace and enum for use with the getKey function.