Make major changes to the way the View handles scrolling
View no longer relies on the response of the Cursor to do scrollling, instead it
looks to see if the Cursor object is outside of the current View (a rectangular
region). If it is, it moves the View and repaints.

When painting, the current View-tangle only is rendered. The repaint function
has been adapted to be based on this instead of more complicated factors.

This change will make the implementation of horizontal scrolling much easier.