Large update which moves to C++0x, see full log
The primary changes contained here are to update some for loops over to C++0x
syntax. Not that big of a change, I know :)

Along with that, though, is a change to make the return type of the ArgParser
callback functions to be void. Nothing is done with the result (for those that
returned one anyway).

One final addition is the util namespace. Currently it just has some simple
wrappers around std functions for searching collections. In the future this
will most likely be expanded.