Add a huge mess that needs straightening, see full log
The idea behind this changeset was to copy an existing "mostly what we want"
configuration system from another of my projects into here, without making
everything get really complicated. We partly succeeded :)

lib/ has been created to house things that really could be used for multiple
projects, so we can keep jail-specific things in src/

lib/config is the new configuration file system, which creates a scoped
variable map sort of thing. The concept behind it was based on what I think
git's config looks like. It may need a complete overhaul, however.

lib/datamap is what backs the config, it is a simple scoped variable map.

Essentially all other changes were to get this to compile. A few functions
were added to util as they are used by the added code. src/config was renamed
src/dconfig, and I'll have to work out how that is going to progress.

lib/config has an argument parsing thing, sort of. It is not nearly as robust
as ArgParser, so it will probably be ditched soonish.

tools/ contain a few programs to work with lib/config, to compile the
defines.cpp file and to dump the default/loaded config file