First round implementation of keymap, see full log
This is not done yet. It isn't even used yet. It does compile, though :)

src/keymap will be the home of a dynamically loaded keymap and dispatcher. It
is passed a Controller/View which it manipulates. Currently the
complie-time-defined functions are defined in here, but that will probably
change very soon into something like src/function. jail only includes the file
to check for compilation issues.

The current minimum set of functions used are default in keymap::init