Return bool from functions to express success, see full log
This also fixes a bug in combine-line. The original keymap was written in a
way that if a function failed somewhere, the rest of it would not execute. So,
for combine-line, if we can't move down we don't want to do the rest of the
steps. Now cursor-down will actually fail and we won't just keep adding
spaces.