Complicate DataMap::load a bit :) see full log
One of the problems encountered in the keymapping portion of the program is
what happens to keys that have multiple names, different cases, abbreviations,
things like that. Basically, the idea implemented is to provide a hook that is
called everytime we're about to change the scope, add a key/value, or we
encounter an error. This function can alter the values at will, and in the
case of an invalid line may even do special parsing and insert the values
directly into the DataMap with the passed DataMap *