Alter the way the function passed to load operates, see full log
Instead of taking several variables, the function now takes a pointer to a
DataMapState structure. This contains the current scope, the line currently
being processed and either the new scope, or new key and value. This way we
have more control over how and what changes. The resulting operation can even
be changed by altering the given type.