Use a non-standard load call to reduce mappings, see full log
We use the newly added functionality of DataMap::load to allow us to inspect
all incoming keys/values to look for things that are sets of keys. Whenever we
find one (key in -mode scope, insert: in value), we take the set of key,
convert them into a vector of tokens, and map them back into a string. This
reduces things like <backspace> and <bspace> into <bspace> and will in the
future handle case insensitivity.